+++ l10n_audit/models.py
@@ -312,7 +312,7 @@ class ResultsRetention:
 # AuditOptions
 # ---------------------------------------------------------------------------
 
-VERSION = "1.2.9"
+VERSION = "1.3.0"
 
 
 @dataclass